+++ src/main/java/dev/enjarai/trickster/spell/mana/SimpleManaPool.java
@@ -1,8 +1,5 @@
 package dev.enjarai.trickster.spell.mana;
 
-import com.mojang.serialization.Codec;
-import com.mojang.serialization.MapCodec;
-import com.mojang.serialization.codecs.RecordCodecBuilder;
 import io.wispforest.endec.Endec;
 import io.wispforest.endec.StructEndec;
 import io.wispforest.endec.impl.StructEndecBuilder;